requirements.txt
pytest
pylint
jupyter
pytest-cov
pytest-flask
pandas
nbval
click
flask
waitress
requests
google-cloud-storage
google-cloud-bigquery
google-cloud-logging
google-cloud-automl
bs4
flasgger
google-api-python-client
locust
dash
dash-bootstrap-components